PromQL: inline the join translation
The vector-match join was laid out by a VectorBinaryOperatorLayout
class of its own. It had one caller, and reading the translator meant
chasing orientation, keys, join and output into another file. Bring it
back as translator methods next to doTranslateBinOpInnerJoin: emitJoin,
emitJoinInput, joinKey, bindJoinOutput, bindJoinResult, reidentify.
No behavior change.
